Split `preconditions` and move to `Attributes`
